api/requirements.txt
requests
bs4
openai
pydantic
langchain
langchain-openai
supabase
streamlit
python-dotenv
fastapi
uvicorn[standard]
apscheduler
elevenlabs
firecrawl-py
sib-api-v3-sdk
pydantic[email]
slack_sdk
PyJWT